Fix C4244 warnings on MSVC 2019

+++ examples/FeatureShowcase/Source/PolylineDrawingTest/PolylineDrawingTest.cpp
@@ -76,8 +76,8 @@ void PolylineDrawingTest::Draw()
     commandList->SetRenderPass(std::move(pass));
 
     auto projectionMatrix = Matrix4x4::CreateOrthographicLH(
-        presentationParameters.BackBufferWidth,
-        presentationParameters.BackBufferHeight,
+        static_cast<float>(presentationParameters.BackBufferWidth),
+        static_cast<float>(presentationParameters.BackBufferHeight),
         0.0f,
         100.0f);
