+++ examples/FeatureShowcase/Source/SpriteLineTest/SpriteLineTest.cpp
@@ -60,8 +60,8 @@ void SpriteLineTest::Draw()
     commandList->Close();
 
     auto projectionMatrix = Matrix4x4::CreateOrthographicLH(
-        presentationParameters.BackBufferWidth,
-        presentationParameters.BackBufferHeight,
+        static_cast<float>(presentationParameters.BackBufferWidth),
+        static_cast<float>(presentationParameters.BackBufferHeight),
         0.0f,
         100.0f);
